Fix DB query

+++ authentication/views.py
@@ -11,42 +11,38 @@ from .serializers import UserSerializer
 from common_utils.get_uid_from_request import get_uid_from_request
 
 
-
 @api_view(['POST', 'DELETE'])
 def AddUser(request):
+    
     if request.method == 'POST':
         # ADD USER TO MATCH FIREBASE UID
         firebase_uid, error_response = get_uid_from_request(request)
-
         if error_response:
             return error_response
         
-        data = {
-            'firebase_uid': firebase_uid
-        }
-
+        data = {'firebase_uid': firebase_uid}
         serializer = UserSerializer(data=data)
         if serializer.is_valid():
             serializer.save()
             return Response({'success': 'Item added successfully'}, status=status.HTTP_201_CREATED)
+        
         return Response({'error': serializer.errors}, status=status.HTTP_400_BAD_REQUEST)
-    
+
     elif request.method == 'DELETE':
         # DELETE USER
         firebase_uid, error_response = get_uid_from_request(request)
-
         if error_response:
             return error_response
 
-        
-        try: 
+        try:
             # Check if user exists
-            delete_user = User.objects.get(firebase_uid=firebase_uid)
+            delete_user = User.objects.filter(firebase_uid=firebase_uid)  
+            if not delete_user.exists():
+                return Response({'error': 'User not found'}, status=status.HTTP_404_NOT_FOUND)
+
             delete_user.delete()
-            return Response({'success': 'User deleted successfully'}, status=status.HTTP_204_NO_CONTENT)
-        except User.DoesNotExist:
-            return Response({'error': 'User not found'}, status=status.HTTP_404_NOT_FOUND)
+            return Response(status=status.HTTP_204_NO_CONTENT)  
+        except Exception as e:
+            return Response({'error': str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
         
-        
-
